fix delete so it can only be done by admin or creator

+++ content-api/resources/methods.py
@@ -99,20 +99,15 @@ def list_subresource(resource_kind, id, subresource_kind):
     )
 
     email = g.verified_email
+    hashed_email = None
+    if email:
+        hashed_email = hashlib.sha256(email.encode("utf-8")).hexdigest()
 
     if auth.user_is_admin(email):
         return json.dumps(matching_children), 200, {"Content-Type": "application/json"}
 
-    if resource_kind == "quizzes" and auth.user_is_creator(email, id):
+    if resource_kind == "quizzes" and auth.user_created_quiz(hashed_email, id):
         return json.dumps(matching_children), 200, {"Content-Type": "application/json"}
-
-    #matching_donors = db.list_matching(
-        #"donors", resource_fields["donors"], "email", email
-    #)
-    #matching_donor_ids = set([donor["id"] for donor in matching_donors])
-    #results = [
-        #item for item in matching_children if item["donor"] in matching_donor_ids
-    #]
 
     return json.dumps(results), 200, {"Content-Type": "application/json"}
 
@@ -158,6 +153,9 @@ def patch(resource_kind, id, representation):
     if resource_kind not in resource_fields:
         return "Not found", 404
 
+    if not auth.allowed("PATCH", resource_kind, representation):
+        return "Forbidden", 403
+
     match_etag = request.headers.get("If-Match", None)
     resource, status = db.update(
         resource_kind, id, representation, resource_fields[resource_kind], match_etag
@@ -180,6 +178,9 @@ def patch(resource_kind, id, representation):
 def delete(resource_kind, id):
     if resource_kind not in resource_fields:
         return "Not found", 404
+
+    if not auth.allowed("DELETE", resource_kind):
+        return "Forbidden", 403
 
     match_etag = request.headers.get("If-Match", None)
     status = db.delete(resource_kind, id, resource_fields[resource_kind], match_etag)
